Refactoring tests and updated android plugin configuration for tests
