+++ common/src/main/java/com/microsoft/identity/common/internal/providers/microsoft/azureactivedirectory/AzureActiveDirectoryOAuth2Strategy.java
@@ -4,11 +4,15 @@ import android.net.Uri;
 
 import com.microsoft.identity.common.Account;
 import com.microsoft.identity.common.internal.net.HttpResponse;
+import com.microsoft.identity.common.internal.net.ObjectMapper;
+import com.microsoft.identity.common.internal.providers.microsoft.MicrosoftTokenErrorResponse;
+import com.microsoft.identity.common.internal.providers.microsoft.MicrosoftTokenResponse;
 import com.microsoft.identity.common.internal.providers.oauth2.AccessToken;
 import com.microsoft.identity.common.internal.providers.oauth2.AuthorizationRequest;
 import com.microsoft.identity.common.internal.providers.oauth2.IDToken;
 import com.microsoft.identity.common.internal.providers.oauth2.OAuth2Strategy;
 import com.microsoft.identity.common.internal.providers.oauth2.RefreshToken;
+import com.microsoft.identity.common.internal.providers.oauth2.TokenErrorResponse;
 import com.microsoft.identity.common.internal.providers.oauth2.TokenRequest;
 import com.microsoft.identity.common.internal.providers.oauth2.TokenResponse;
 import com.microsoft.identity.common.internal.providers.oauth2.TokenResult;
@@ -116,7 +120,17 @@ public class AzureActiveDirectoryOAuth2Strategy extends OAuth2Strategy {
 
     @Override
     protected TokenResult getTokenResultFromHttpResponse(HttpResponse response){
-        return null;
+        TokenResponse tokenResponse = null;
+        TokenErrorResponse tokenErrorResponse = null;
+
+        if(response.getStatusCode() >= 400){
+            //An error occurred
+            tokenErrorResponse = (TokenErrorResponse) ObjectMapper.deserializeJsonStringToObject(response.getBody(), MicrosoftTokenErrorResponse.class);
+        }else{
+            tokenResponse = (TokenResponse)ObjectMapper.deserializeJsonStringToObject(response.getBody(), AzureActiveDirectoryTokenResponse.class);
+        }
+
+        return new TokenResult(tokenResponse, tokenErrorResponse);
     }
 
 }
