+++ common/src/test/java/com/microsoft/identity/common/unit/ObjectMapperTest.java
@@ -1,4 +1,4 @@
-package com.microsoft.identity.common;
+package com.microsoft.identity.common.unit;
 
 
 import com.microsoft.identity.common.internal.net.ObjectMapper;
